+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project will adhere to [Semantic Versioning](https://semver.org/spec/v2
 
 - Circle.circumference
 - Circle.area
+- Circle.collidePoint
 
 ## [0.0.8] - 2024-01-08